Add base-with-prettier preset as a prettier alternative
Extract the formatter-agnostic part of the base config into a new
configs/base-shared.js module, then have configs/base.js wrap it with
dprint (default) and a new configs/base-with-prettier.js wrap it with
prettier instead. Each is a complete base preset; consumers pick one
or the other but never combine them.

The new preset is published as @enormora/eslint-config-base-with-prettier
and ships its own readme. Update the presets overview in the project
readme and reword base.md to clarify dprint is the default formatter.
