+++ package.json
@@ -29,6 +29,7 @@
         "eslint-plugin-no-barrel-files": "1.3.1",
         "eslint-plugin-no-secrets": "2.3.3",
         "eslint-plugin-perfectionist": "5.9.0",
+        "eslint-plugin-prettier": "5.5.5",
         "eslint-plugin-promise": "7.3.0",
         "eslint-plugin-react": "7.37.5",
         "eslint-plugin-react-hooks": "7.0.1",
